Move ContactsViewModel
